Andreev-Kingkade a0 (mean age at death in the first year of life)

Period (m0-based) formulas of Andreev and Kingkade (2015) as tabulated in HMD Methods Protocol V6, Table 3. Each interval is closed at its lower bound.

andreev_kingkade_a0(m0, sex = c("female", "male"))

Arguments

m0

Central death rate at age 0 (a single number).

sex

"female" or "male" (case-insensitive).

Value

Numeric a0 value, NA if m0 is NA.

Details

Sex	m0 range	a0

Male	[0, 0.02300)	0.14929 - 1.99545 m0

Male	[0.02300, 0.08307)	0.02832 + 3.26021 m0

Male	[0.08307, Inf)	0.29915

Female	[0, 0.01724)	0.14903 - 2.05527 m0

Female	[0.01724, 0.06891)	0.04667 + 3.88089 m0

Female	[0.06891, Inf)	0.31411

References

Andreev, E. M. and Kingkade, W. W. (2015). Average age at death in infancy and infant mortality level: Reconsidering the Coale-Demeny formulas at current levels of low mortality. Demographic Research 33(13): 363-390. doi:10.4054/DemRes.2015.33.13

Examples

andreev_kingkade_a0(0.004, "female")
#> [1] 0.1408089
andreev_kingkade_a0(0.15, "male")
#> [1] 0.29915